refactor: improve removing messages sent before the clearedAt timestamp into setRooms setter
