+++ src/types/store/RoomTypes.ts
@@ -4,6 +4,33 @@
  * SPDX-License-Identifier: AGPL-3.0-only
  */
 
+import { MemberBe, RoomBe } from '../network/models/roomBeTypes';
+
+export type RoomsStoreSlice = {
+	rooms: RoomsMap;
+	setRooms: (rooms: RoomBe[]) => void;
+	addRoom: (room: RoomBe) => void;
+	deleteRoom: (id: string) => void;
+	setRoomName: (id: string, newName: string) => void;
+	setRoomDescription: (id: string, newDescription: string) => void;
+	setRoomNameAndDescription: (
+		id: string,
+		newName: string | undefined,
+		newDescription: string | undefined
+	) => void;
+	setRoomMuted: (id: string) => void;
+	setRoomUnmuted: (id: string) => void;
+	addRoomMember: (id: string, member: MemberBe) => void;
+	removeRoomMember: (id: string, userId: string | undefined) => void;
+	promoteMemberToModerator: (id: string, userId: string) => void;
+	demoteMemberFromModerator: (id: string, userId: string) => void;
+	setClearedAt: (roomId: string, clearedAt: string) => void;
+	setRoomPictureUpdated: (id: string, date: string) => void;
+	setRoomPictureDeleted: (id: string) => void;
+	setPlaceholderRoom: (userId: string) => void;
+	replacePlaceholderRoom: (userId: string, newRoomId: string) => void;
+};
+
 export type Room = {
 	id: string;
 	name?: string;
